Added guard to check if current user created the post in order to delete

+++ src/components/reviews/reviewIDetail/ReviewDetail.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import { Link, useSubmit } from 'react-router-dom';
+import { RootState } from '../../../store';
+import { useSelector } from 'react-redux';
 import { ReviewItem as IReviewItem } from '../../../pages/reviews/Reviews';
 import './reviewDetail.scss';
 
@@ -7,12 +9,17 @@ interface ReviewDetailProps {
   review: IReviewItem;
 }
 
+// TODO Make modal for this proceed window
+
 const ReviewDetail: React.FC<ReviewDetailProps> = ({ review }) => {
   const submit = useSubmit();
+  const uid = useSelector<RootState>(state => state.auth.currentUser.uid);
+
+  const usersPost = uid === review.uid && uid !== undefined;
 
   const deleteEventHandler = () => {
     const proceed = window.confirm('Are you sure?');
-    if (proceed) {
+    if (proceed && usersPost) {
       submit(null, { method: 'delete' });
     }
   };
@@ -24,7 +31,7 @@ const ReviewDetail: React.FC<ReviewDetailProps> = ({ review }) => {
       <p>{review.userName}</p>
       <p>{review.description}</p>
       <Link to="..">Back to Reviews</Link>
-      <button onClick={deleteEventHandler}>Delete</button>
+      {usersPost && <button onClick={deleteEventHandler}>Delete</button>}
     </article>
   );
 };
